[TRIVIAL] Use `mod.rs` consistently and clean up a few tiny files
