+++ crates/driver/src/boundary/liquidity/swapr.rs
@@ -5,10 +5,7 @@ use {
         infra::{self, blockchain::Ethereum},
     },
     ethrpc::block_stream::CurrentBlockWatcher,
-    shared::sources::{
-        swapr::reader::SwaprPoolReader,
-        uniswap_v2::pool_fetching::DefaultPoolReader,
-    },
+    shared::sources::{swapr::SwaprPoolReader, uniswap_v2::pool_fetching::DefaultPoolReader},
     solver::{liquidity::ConstantProductOrder, liquidity_collector::LiquidityCollecting},
 };
