+++ crates/shared/src/trade_finding/mod.rs
