+++ crates/solver/src/interactions/mod.rs
