+++ crates/solver/src/solver/mod.rs
